+++ Changelog.md
@@ -2,12 +2,18 @@
 
 #### Consider supporting me on [Patreon](https://patreon.com/mooshi69) or [Ko-Fi](https://ko-fi.com/mooshi69)!
 
+## // Decemper 11th 2024
+
+- Made the search method for the scanlators optional
+- Added the 'supports_search' property in the schema file
+
 ## // November 25th 2024
 
 ### Bug Fixes:
 
 - Fixed flamecomics website
 - Added support for https://www.beyondtheataraxia.com/
+- Added support for https://www.gyarelease.it/
 - Fixed asura chapter links
 
 ## // November 12th 2024